minor fixes

+++ Backend/src/main/java/it/unicam/cs/ids/loyaltyPlatform/model/campaign/CampaignServiceImpl.java
@@ -1,6 +1,5 @@
 package it.unicam.cs.ids.loyaltyPlatform.model.campaign;
 
-import jakarta.persistence.Table;
 import lombok.NonNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,7 +10,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-@Table(name = "")
 public class CampaignServiceImpl implements CampaignService {
 
     @Autowired
@@ -19,52 +17,22 @@ public class CampaignServiceImpl implements CampaignService {
 
     @Override
     public Campaign save(@NonNull Campaign campaign) {
-        return this.repository.save(campaign);
-    }
-
-    @Override
-    public void saveById(@NonNull Long id) {
-        //TODO implement
-        // repository doesn't provide a ready to use saveById method.
-    }
-
-    @Override
-    public Campaign update(@NonNull Campaign campaign) {
-        Campaign oldCampaign = this.repository.findById(campaign.getId()).orElse(null);
-
-        if (existsById(oldCampaign.getId())) {
-            if (oldCampaign.getId().equals(campaign.getId())) {
-                return this.repository.save(campaign);
-            } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "...");
-        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "non-existent campaign.");
-    }
-
-    @Override
-    public Campaign updateById(@NonNull Long id) {
-        Optional<Campaign> old = this.repository.findById(id);
-
-        if (old.isPresent()) {
-            Campaign originalCampaign = old.get();
-            return this.repository.save(originalCampaign);
-        }
-        return null;
+        if (!this.repository.existsById(campaign.getId())) {
+            this.repository.save(campaign);
+            return campaign;
+        } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Campaign is already present!");
     }
 
     @Override
     public Campaign findById(@NonNull Long id) {
-        //one of the lines below should be removed
-        //return this.repository.getReferenceById(id);
         return this.repository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found"));
     }
 
     @Override
     public Optional<Campaign> getById(@NonNull Long id) {
-        Optional<Campaign> campaign = this.repository.findById(id);
-        if (campaign.isPresent()) {
+        if (this.repository.findById(id).isPresent()) {
             return this.repository.findById(id);
-        } else {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found");
-        }
+        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found!");
     }
 
     @Override
@@ -73,8 +41,28 @@ public class CampaignServiceImpl implements CampaignService {
     }
 
     @Override
+    public Campaign update(@NonNull Campaign campaign) {
+        if (this.repository.existsById(campaign.getId())) {
+            this.repository.save(campaign);
+            return campaign;
+        } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Campaign does not exist!");
+    }
+
+    @Override
+    public Campaign updateById(@NonNull Long id) {
+        if (this.repository.existsById(id)) {
+            this.repository.save(findById(id));
+            return findById(id);
+        } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Trying to update non-existing campaign!");
+    }
+
+    @Override
     public void delete(@NonNull Campaign campaign) {
-        this.repository.delete(campaign);
+        for (Campaign c : this.repository.findAll()) {
+            if (this.repository.existsById(c.getId())) {
+                this.repository.delete(campaign);
+            } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found");
+        }
     }
 
     @Override
@@ -82,13 +70,7 @@ public class CampaignServiceImpl implements CampaignService {
         if (this.repository.findById(id).isPresent()) {
             this.repository.deleteById(id);
             System.out.println("Campaign deleted successfully");
-        } else {
-            System.out.println("No such campaign in the database");
-        }
-    }
-
-    private boolean existsById(@NonNull Long id) {
-        return this.repository.existsById(id);
+        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found");
     }
 
 }
