+++ Backend/src/main/java/it/unicam/cs/ids/loyaltyPlatform/model/company/CompanyServiceImpl.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CompanyServiceImpl implements CompanyService {
@@ -16,12 +15,17 @@ public class CompanyServiceImpl implements CompanyService {
     private CompanyRepository repository;
 
     public Company save(@NonNull Company company) {
-        return this.repository.save(company);
+        if (!this.repository.existsById(company.getId())) {
+            this.repository.save(company);
+            return company;
+        } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Company is already present!");
     }
 
     @Override
     public Company saveById(@NonNull Long id) {
-        return this.repository.save(this.findById(id));
+        if (!this.repository.existsById(id)) {
+            return this.repository.save(this.findById(id));
+        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Company already exists!");
     }
 
     @Override
@@ -36,30 +40,35 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public Company update(@NonNull Company company) {
-        //TODO implement business logic here
-        return this.repository.save(company);
+        if (this.repository.existsById(company.getId())) {
+            this.repository.save(company);
+            return company;
+        } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Campaign does not exist!");
     }
 
     @Override
     public Company updateById(@NonNull Long id) {
-
-        Optional<Company> oldCompany = this.repository.findById(id);
-
-        if (oldCompany.isPresent()) {
-            Company orgCompany = oldCompany.get();
-            return this.repository.save(orgCompany);
-        }
-        return null;
+        if (this.repository.existsById(id)) {
+            this.repository.save(findById(id));
+            return findById(id);
+        } else throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Trying to update non-existing campaign!");
     }
 
     @Override
     public void delete(@NonNull Company company) {
-        this.repository.delete(company);
+        for (Company c : this.repository.findAll()) {
+            if (this.repository.existsById(c.getId())) {
+                this.repository.delete(company);
+            } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found");
+        }
     }
 
     @Override
     public void deleteById(@NonNull Long id) {
-        this.repository.deleteById(id);
+        if (this.repository.findById(id).isPresent()) {
+            this.repository.deleteById(id);
+            System.out.println("Campaign deleted successfully");
+        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Campaign not found");
     }
 
 }
